perception/webapi/requirements.txt
flask
flask-cors
numpy
opencv-python
torch
trimesh
scipy
Pillow
requests
imageio
